@OneToMany를 15년 후에 죽여주세요

+++ src/main/kotlin/com/solve/domain/post/dto/response/PostCommentReplyResponse.kt
@@ -8,7 +8,7 @@ data class PostCommentReplyResponse(
     val id: Long,
     val content: String,
     val author: PostCommentReplyAuthorResponse,
-    val likeCount: Int,
+    val likeCount: Long,
     val liked: Boolean,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime
