+++ src/main/kotlin/com/solve/domain/post/dto/response/PostCommentResponse.kt
@@ -8,7 +8,7 @@ data class PostCommentResponse(
     val id: Long,
     val content: String,
     val author: PostCommentAuthorResponse,
-    val likeCount: Int,
+    val likeCount: Long,
     val liked: Boolean,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime
